[TASK] Drop `AbstractUserController`
Now that there is only one controller, we do not need the abstract base class
anymore.
